added handbook package

+++ src/main/java/com/cdek/java/model/request/Service.java
@@ -1,14 +1,16 @@
 package com.cdek.java.model.request;
 
-import java.math.BigDecimal;
 import lombok.Getter;
 import lombok.Setter;
 
+/**
+ * Данные о дополнительных услугах
+ */
 @Getter
 @Setter
 public class Service {
 
   private Integer code;
-  private BigDecimal parameter;
+  private Integer parameter;
 
 }
